requirements.txt
pycryptodomex==3.18.0
PyQt5==5.15.9
PyQt5_sip==12.12.1
rsa==4.9
